+++ src/main/java/com/popjub/reservationservice/application/valid/ReservationValidator.java
@@ -9,8 +9,8 @@ import com.popjub.reservationservice.application.dto.command.CreateReservationCo
 import com.popjub.reservationservice.application.service.NoShowService;
 import com.popjub.reservationservice.domain.entity.ReservationStatus;
 import com.popjub.reservationservice.domain.repository.ReservationRepository;
-import com.popjub.reservationservice.exception.ReservationCustomException;
-import com.popjub.reservationservice.exception.ReservationErrorCode;
+import com.popjub.reservationservice.global.exception.ReservationCustomException;
+import com.popjub.reservationservice.global.exception.ReservationErrorCode;
 
 import lombok.RequiredArgsConstructor;
